Separate auth hook from context module
