+++ webui/src/app/providers/auth-context.ts
@@ -1,4 +1,4 @@
-import {createContext, useContext} from "react";
+import {createContext} from "react";
 import type {CurrentUser} from "../../shared/api/auth";
 
 export type AuthStatus = "loading" | "authenticated" | "unauthenticated";
@@ -11,11 +11,3 @@ export type AuthContextValue = {
 };
 
 export const AuthContext = createContext<AuthContextValue | null>(null);
-
-export function useAuth(): AuthContextValue {
-  const value = useContext(AuthContext);
-  if (value === null) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return value;
-}
